Consider translating nil
If we gave nil to translate, an exception would be thrown. This patches
changes the `t` fn so that, if it gets nil, just returns nil. Seems an
appropriate scenario. If we don't have nothing to translate, return
nothing.

This became necessary on the reports, to translate fields. Sometimes the
user fields are nil. This way we can just pass the field value.